Check if cache file exists when reset

+++ bin/pythoncompile.py
@@ -1188,7 +1188,8 @@ if __name__ == '__main__':
             elif o in ("-init","--initialize"):
                 initialize = True
             elif o in ("-R","--reset"):
-                os.remove(".cache_metacompile.json");
+                if(os.path.isfile('.cache_metacompile.json')):
+                    os.remove(".cache_metacompile.json");
 
     except getopt.GetoptError as err:
         LOG.error(err)
